CORE-9072 added more error checking in run script

+++ testing/run.sh
@@ -3,27 +3,40 @@
 
 main()
 {
+  local baseDir=$(dirname $(readlink -f "$0"))
+
   if [ "$#" -lt 1 ]
   then
-    printf 'The name of a folder holding the playbook is requred\n'
+    printf 'The name of a folder holding the playbook is requred\n' >&2
     return 1
   fi
 
-  local baseDir=$(dirname $(readlink -f "$0"))
   local playbooks="$1"
-
-  if [ "$#" -ge 2 ]
-  then
-    playbook="$2"
-  else
-    playbook=main.yml
-  fi
 
   if ! [[ "$playbooks" =~ ^/ ]]
   then
     playbooks="$(pwd)"/"$playbooks"
   fi
 
+  if ! [ -d "$playbooks" ]
+  then
+    printf '%s is not a directory\n' "$playbooks" >&2
+    return 1
+  fi
+
+  if [ "$#" -ge 2 ]
+  then
+    local playbook="$2"
+  else
+    local playbook=main.yml
+  fi
+
+  if ! [ -f "$playbooks"/"$playbook" ]
+  then
+    printf 'The playbook %s/%s does not exist\n' "$playbooks" "$playbook" >&2
+    return 1
+  fi
+  
   if "$baseDir"/env/controller start
   then
     docker run --rm --tty \
